requirements.txt
evdev
python-mpd2 
pytest-mock
validators